+++ StoreRotationConfig/Patches/TerminalItemSalesPatches.cs
@@ -188,7 +188,7 @@ namespace StoreRotationConfig.Patches
                         TerminalNode item = storeRotation[index];
 
                         // Return if 'salesChance' is disabled OR the 'RotationSales' dictionary doesn't contain a discount for the item about to be displayed.
-                        if (Plugin.Settings.SALE_CHANCE == 0 || RotationSales == null || !RotationSales.ContainsKey(item))
+                        if (Plugin.Settings.SALE_CHANCE == 0 || (RotationSales != null && !RotationSales.ContainsKey(item)))
                         {
                             return;
                         }